+++ src/styles/scss/ChatbotPage.scss
